// src/app/models/UserDetails.ts
export class UserDetails {
    UserID: number;
    FirstName: string;
    LastName: string;
    FullName: string;
    EmailID: string;
    PhoneNo: string;
    Address: string;
    Area: string;
    City: string;
    State: string;
    Country: string;
    Gender: string;
    MemberShipType: string;
    MembershipExpiry: string;

    DateOfBirth: Date;

    MarriageDate: Date;

    IsActive: boolean;

    UserReferanceCode: string;

    ReferedByReferanceCode: string;

    BalanceMessages: number;

    UpiId: string;

    AccountNo: string;

    IFSCCode: string;

    AreaID: number;

    CityID: number;

    StateID: number;

    CountryID: number;

    GenderID: number;

    MembershipTypeID: number;
}